+++ dev/test/detailedResponse_tests.ts
@@ -33,17 +33,17 @@ const wrappedDetailError = (
   getDetailedError({
     success,
     comparator: comparator || "==",
-    getOperandDisplayNames: (item) => ({
-      operand1Name: item.operand1Name,
-      operand2Name: item.operand2Name,
+    getOperandDisplayNames: (item: TestItem) => ({
+      actualName: item.operand1Name,
+      expectedName: item.operand2Name,
     }),
-    getOperandIds: (item) => ({
-      operand1Id: item.operand1Id,
-      operand2Id: item.operand2Id,
+    getOperandIds: (item: TestItem) => ({
+      actualId: item.operand1Id,
+      expectedId: item.operand2Id,
     }),
     item,
-    operand1: item.operand1Value,
-    operand2: item.operand2Value,
+    actual: item.operand1Value,
+    expected: item.operand2Value,
   });
 
 describe("Detailed Response", () => {
@@ -73,8 +73,8 @@ describe("Detailed Response", () => {
     it("shall contain 'NOT' in the message if negative is true ", () => {
       const msg = getHumanReadableMessage({
         comparator: "==",
-        operand1Name: "First",
-        operand2Name: "Second",
+        actualName: "First",
+        expectedName: "Second",
         useNegative: true,
       });
       expect(msg).to.contain("NOT");
@@ -83,8 +83,8 @@ describe("Detailed Response", () => {
     it("shall NOT contain the word 'NOT' in the message if negative is falsy ", () => {
       const msg = getHumanReadableMessage({
         comparator: "==",
-        operand1Name: "First",
-        operand2Name: "Second",
+        actualName: "First",
+        expectedName: "Second",
         useNegative: false,
       });
       expect(msg).to.not.contain("NOT");
@@ -93,8 +93,8 @@ describe("Detailed Response", () => {
     it("shall allow the user to override the translated operator", () => {
       const msg = getHumanReadableMessage({
         comparator: "==",
-        operand1Name: "First",
-        operand2Name: "Second",
+        actualName: "First",
+        expectedName: "Second",
         overrideTranslation: () => "TEST OVERRIDE",
       });
       expect(msg).to.contain("TEST OVERRIDE");