+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 - 2020-1-3
+
+Break out comparator functionality for external testing and change operand names to
+actual and expected
+
 ## 1.0.0 - 2020-12-31
 
 Initial Release